refactor: web でしか使わないコードを common から web へ移動
common は全パッケージ共通基盤だが、実際には web からのみ参照される
コードが含まれていたため web 配下へ移す。

- errors/handleError: Result のエラーを Hono の HTTPException へ変換する
  サーバー/HTTP 専用処理。web の server 層でしか使わないため
  apps/web/src/server/handle-error.ts へ移動
- locale の表示整形系（toJaDateString / toJaTimeString /
  toTodayJstDateString / formatSummaryDateTick）: UI 表示・チャート軸用で
  web からのみ利用するため apps/web/src/common/locale/date.ts へ移動。
  domain 等と共有する addJstDays / toJstDate / toJstDateString は common に残す

なお pagination の MAX_PAGE は共有スキーマ offsetPaginationSchema が内部で
参照しており、web 専用ではないため common に残した。
