+++ application/apps/web/src/server/article/handler/get-articles.ts
@@ -1,4 +1,3 @@
-import { handleError } from '@trend-diary/common/errors'
 import type { OffsetPaginationResult } from '@trend-diary/common/pagination'
 import { offsetPaginationSchema } from '@trend-diary/common/pagination'
 import getRdbClient from '@trend-diary/datastore/rdb'
@@ -12,6 +11,7 @@ import {
 import { z } from 'zod'
 import CONTEXT_KEY from '@/middleware/context'
 import type { ZodValidatedQueryContext } from '@/middleware/zod-validator'
+import { handleError } from '@/server/handle-error'
 import { type ArticleResponse, toArticleResponse } from '../article-response'
 
 const readStatusEnum = z.enum(['0', '1'])
