+++ application/apps/web/src/server/auth/handler/logout.ts
@@ -1,7 +1,7 @@
 import { authClientConfig, PasswordAuthClient } from '@trend-diary/authentication'
-import { handleError } from '@trend-diary/common/errors'
 import type { Context } from 'hono'
 import CONTEXT_KEY from '@/middleware/context'
+import { handleError } from '@/server/handle-error'
 import toAuthError from '../auth-error'
 
 export default async function logout(c: Context) {
