+++ application/apps/web/src/server/auth/handler/passkey-login-start.ts
@@ -1,8 +1,8 @@
 import { authClientConfig, PasskeyClient } from '@trend-diary/authentication'
-import { handleError } from '@trend-diary/common/errors'
 import type { Context } from 'hono'
 import type { Env } from '@/env'
 import CONTEXT_KEY from '@/middleware/context'
+import { handleError } from '@/server/handle-error'
 import toAuthError from '../auth-error'
 
 export default async function passkeyLoginStart(c: Context<Env>) {
