+++ application/apps/web/src/server/auth/handler/passkey-register-verify.ts
@@ -1,9 +1,9 @@
 import type { RegistrationResponseJSON } from '@simplewebauthn/browser'
 import { authClientConfig, PasskeyClient } from '@trend-diary/authentication'
-import { handleError } from '@trend-diary/common/errors'
 import { z } from 'zod'
 import CONTEXT_KEY from '@/middleware/context'
 import type { ZodValidatedContext } from '@/middleware/zod-validator'
+import { handleError } from '@/server/handle-error'
 import toAuthError from '../auth-error'
 
 // 真正性はSupabaseが検証するため中身の妥当性検証はプロバイダに委ね、ここは登録 ceremony 結果を素通しする
