+++ application/apps/web/src/server/handler-factory.ts
@@ -54,7 +54,6 @@
  * - deletePermission.ts（動的パス + $delete）
  */
 
-import { handleError } from '@trend-diary/common/errors'
 import type { LoggerType } from '@trend-diary/common/logger'
 import getRdbClient, { type RdbClient } from '@trend-diary/datastore/rdb'
 import type { Context } from 'hono'
@@ -63,6 +62,7 @@ import type { ContentfulStatusCode, StatusCode } from 'hono/utils/http-status'
 import { type Result } from 'neverthrow'
 import type { Env, SessionUser } from '@/env'
 import CONTEXT_KEY from '@/middleware/context'
+import { handleError } from '@/server/handle-error'
 
 // コンテキストの型定義
 // oxlint-disable-next-line typescript/no-restricted-types -- ハンドラ側で具象型を指定するまでの既定値で、任意形状を受け入れる必要があるため
