+++ application/packages/common/src/errors/index.ts
@@ -2,14 +2,6 @@ import AlreadyExistsError from './client-error/already-exists-error'
 import ClientError from './client-error/client-error'
 import NotFoundError from './client-error/not-found-error'
 import ExternalServiceError from './external-service-error'
-import handleError from './handle'
 import ServerError from './server-error'
 
-export {
-  AlreadyExistsError,
-  ClientError,
-  ExternalServiceError,
-  handleError,
-  NotFoundError,
-  ServerError,
-}
+export { AlreadyExistsError, ClientError, ExternalServiceError, NotFoundError, ServerError }
